Try to infer the type when suffix is omitted
Instead of giving up, try to be a bit smarter and infer the non-zero
type. Unfortunately, Rust won't infer the type for us like it does for
integer literals, but we can hack a thing that works most of the time
with from/Into conversions. This covers the cases when the type is
obvious, such as a struct field, a return value, a parameter.